+++ packages/dart_firebase_admin/lib/src/google_cloud_firestore/reference.dart
@@ -3,14 +3,13 @@ part of 'firestore.dart';
 final class CollectionReference<T> extends Query<T> {
   CollectionReference._({
     required super.firestore,
-    required _QualifiedResourcePath path,
+    required _ResourcePath path,
     required _FirestoreDataConverter<T> converter,
   }) : super._(
           queryOptions: _QueryOptions.forCollectionQuery(path, converter),
         );
 
-  _QualifiedResourcePath get _resourcePath =>
-      _queryOptions.parentPath._append(id) as _QualifiedResourcePath;
+  _ResourcePath get _resourcePath => _queryOptions.parentPath._append(id);
 
   /// The last path element of the referenced collection.
   String get id => _queryOptions.collectionId;
@@ -28,7 +27,7 @@ final class CollectionReference<T> extends Query<T> {
 
     return DocumentReference<DocumentData>._(
       firestore: firestore,
-      path: _queryOptions.parentPath as _QualifiedResourcePath,
+      path: _queryOptions.parentPath,
       converter: _jsonConverter,
     );
   }
@@ -62,7 +61,7 @@ final class CollectionReference<T> extends Query<T> {
     }
 
     if (!identical(_queryOptions.converter, _jsonConverter) &&
-        path._parent() != _resourcePath) {
+        path.parent() != _resourcePath) {
       throw ArgumentError.value(
         documentPath,
         'documentPath',
@@ -159,12 +158,12 @@ final class CollectionReference<T> extends Query<T> {
 final class DocumentReference<T> implements _Serializable {
   const DocumentReference._({
     required this.firestore,
-    required _QualifiedResourcePath path,
+    required _ResourcePath path,
     required _FirestoreDataConverter<T> converter,
   })  : _converter = converter,
         _path = path;
 
-  final _QualifiedResourcePath _path;
+  final _ResourcePath _path;
   final _FirestoreDataConverter<T> _converter;
   final Firestore firestore;
 
@@ -187,7 +186,7 @@ final class DocumentReference<T> implements _Serializable {
   CollectionReference<T> get parent {
     return CollectionReference<T>._(
       firestore: firestore,
-      path: _path._parent()!,
+      path: _path.parent()!,
       converter: _converter,
     );
   }
@@ -200,6 +199,40 @@ final class DocumentReference<T> implements _Serializable {
           firestore._databaseId,
         )
         ._formattedName;
+  }
+
+  /// Fetches the subcollections that are direct children of this document.
+  ///
+  /// ```dart
+  /// final documentRef = firestore.doc('col/doc');
+  ///
+  /// documentRef.listCollections().then((collections) {
+  ///   for (final collection in collections) {
+  ///     print('Found subcollection with id: ${collection.id}');
+  ///   }
+  /// });
+  /// ```
+  Future<List<CollectionReference<DocumentData>>> listCollections() {
+    return this.firestore._client.v1((a) async {
+      final request = firestore1.ListCollectionIdsRequest(
+        // Setting `pageSize` to an arbitrarily large value lets the backend cap
+        // the page size (currently to 300). Note that the backend rejects
+        // MAX_INT32 (b/146883794).
+        pageSize: (math.pow(2, 16) - 1).toInt(),
+      );
+
+      final result = await a.projects.databases.documents.listCollectionIds(
+        request,
+        this._formattedName,
+      );
+
+      final ids = result.collectionIds ?? [];
+      ids.sort((a, b) => a.compareTo(b));
+
+      return [
+        for (final id in ids) collection(id),
+      ];
+    });
   }
 
   /// Changes the de/serializing mechanism for this [DocumentReference].
@@ -416,10 +449,6 @@ class _QueryCursor {
 
   @override
   bool operator ==(Object other) {
-    // if (other is! _QueryCursor) return false;
-
-    // print(_valuesEqual(values, other.values));
-
     return other is _QueryCursor &&
         runtimeType == other.runtimeType &&
         before == other.before &&
@@ -510,11 +539,11 @@ class _QueryOptions<T> with _$QueryOptions<T> {
 
   /// Returns query options for a single-collection query.
   factory _QueryOptions.forCollectionQuery(
-    _QualifiedResourcePath collectionRef,
+    _ResourcePath collectionRef,
     _FirestoreDataConverter<T> converter,
   ) {
     return _QueryOptions<T>(
-      parentPath: collectionRef._parent()!,
+      parentPath: collectionRef.parent()!,
       collectionId: collectionRef.id!,
       converter: converter,
       allDescendants: false,