+++ packages/dart_firebase_admin/test/google_cloud_firestore/document_test.dart
@@ -9,8 +9,30 @@ void main() {
     late DocumentReference<Map<String, Object?>> documentRef;
 
     setUp(() {
-      firestore = createInstance();
+      firestore = createFirestore();
       documentRef = firestore.doc('collectionId/documentId');
+    });
+
+    test('listCollections', () async {
+      final doc1 = firestore.doc('collectionId/a');
+      final doc2 = firestore.doc('collectionId/b');
+
+      final doc1col1 = doc1.collection('a');
+      final doc1col2 = doc1.collection('b');
+
+      final doc2col1 = doc2.collection('c');
+      final doc2col2 = doc2.collection('d');
+
+      await doc1col1.add({});
+      await doc1col2.add({});
+      await doc2col1.add({});
+      await doc2col2.add({});
+
+      final doc1Collections = await doc1.listCollections();
+      final doc2Collections = await doc2.listCollections();
+
+      expect(doc1Collections, unorderedEquals([doc1col1, doc1col2]));
+      expect(doc2Collections, unorderedEquals([doc2col1, doc2col2]));
     });
 
     test('has collection() method', () {
@@ -60,7 +82,7 @@ void main() {
   group('serialize document', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test("doesn't serialize unsupported types", () {
       expect(
@@ -98,7 +120,7 @@ void main() {
     });
 
     test('Supports BigInt', () async {
-      final firestore = createInstance(Settings(useBigInt: true));
+      final firestore = createFirestore(Settings(useBigInt: true));
 
       await firestore.doc('collectionId/bigInt').set({
         'foo': BigInt.from(9223372036854775807),
@@ -199,10 +221,10 @@ void main() {
   group('get document', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test('returns document', () async {
-      firestore = createInstance();
+      firestore = createFirestore();
       await firestore.doc('collectionId/getdocument').set({
         'foo': {
           'bar': 'foobar',
@@ -280,7 +302,7 @@ void main() {
   group('delete document', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test('works', () async {
       await firestore.doc('collectionId/deletedoc').set({});
@@ -333,7 +355,7 @@ void main() {
   group('set documents', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test('sends empty non-merge write even with just field transform',
         () async {
@@ -386,7 +408,7 @@ void main() {
   group('create document', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test('creates document', () async {
       await firestore.doc('collectionId/createdoc').delete();
@@ -433,7 +455,7 @@ void main() {
   group('update document', () {
     late Firestore firestore;
 
-    setUp(() => firestore = createInstance());
+    setUp(() => firestore = createFirestore());
 
     test('works', () async {
       await firestore.doc('collectionId/updatedoc').set({'foo': 'bar'});